# specs/009-investment-workbench/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Lucking 投资工作台 API
  version: 1.0.0
  description: 前端唯一公共契约；所有端点均为同域 /api/v1 路径。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /auth/login:
    post:
      security: []
      summary: 账号密码登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功并设置 HttpOnly 会话 Cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      summary: 退出当前会话
      operationId: logout
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: 当前会话已删除并清除 Cookie
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/me:
    get:
      summary: 获取当前用户和 CSRF Token
      operationId: getCurrentUser
      responses:
        '200':
          description: 当前会话
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/password:
    put:
      summary: 修改当前用户密码并撤销全部会话
      operationId: changePassword
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: 密码修改成功，全部会话已撤销
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /calendar:
    get:
      summary: 获取交易日历和当前用户重要日
      operationId: listCalendar
      parameters:
        - name: start_date
          in: query
          required: true
          schema: { type: string, format: date }
        - name: end_date
          in: query
          required: true
          schema: { type: string, format: date }
        - name: market_code
          in: query
          schema: { type: string, enum: [CN-S], default: CN-S }
      responses:
        '200':
          description: 日期范围结果；交易日缺失必须返回 UNKNOWN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /important-dates:
    post:
      summary: 新增当前用户重要日
      operationId: createImportantDate
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportantDateInput' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportantDateResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /important-dates/{important_date_id}:
    put:
      summary: 修改当前用户重要日
      operationId: updateImportantDate
      parameters:
        - $ref: '#/components/parameters/ImportantDateId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportantDateInput' }
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportantDateResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: 删除当前用户重要日
      operationId: deleteImportantDate
      parameters:
        - $ref: '#/components/parameters/ImportantDateId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { description: 删除成功 }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /broker-recommendations:
    get:
      summary: 查询券商金股
      operationId: listBrokerRecommendations
      parameters:
        - name: recommendation_month
          in: query
          required: true
          description: 推荐月份，以当月首日表示
          schema: { type: string, format: date, pattern: '^\d{4}-\d{2}-01$' }
        - name: broker_name
          in: query
          schema: { type: string, maxLength: 160 }
        - name: stock_id
          in: query
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: 券商金股分页结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrokerRecommendationListResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: lucking_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 32, maxLength: 256 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    ImportantDateId:
      name: important_date_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: 请求字段或业务范围无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/BadRequestErrorResponse' }
    Unauthorized:
      description: 未认证或会话已失效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UnauthorizedErrorResponse' }
    Forbidden:
      description: CSRF、同源或资源所有权校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ForbiddenErrorResponse' }
    NotFound:
      description: 资源不存在或不属于当前用户
      content:
        application/json:
          schema: { $ref: '#/components/schemas/NotFoundErrorResponse' }
    Conflict:
      description: 唯一性或并发版本冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ConflictErrorResponse' }
    RateLimited:
      description: 请求频率受限
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RateLimitedErrorResponse' }
...
    InternalServerError:
      description: 未预期服务错误；不得返回堆栈、SQL、秘密或第三方原始响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/InternalServerErrorResponse' }
  schemas:
    BusinessErrorCode:
      type: integer
      description: 稳定业务错误码；HTTP 状态表达协议结果，业务码表达应用原因
      enum:
        - 100001
        - 100002
        - 100003
        - 200001
        - 200002
        - 200003
        - 300001
        - 400001
        - 400002
        - 400003
        - 400004
        - 500001
        - 500002
        - 900001
      x-enum-varnames:
        - INVALID_CREDENTIALS
        - SESSION_INVALID
        - CSRF_VALIDATION_FAILED
        - REQUEST_VALIDATION_FAILED
        - QUERY_RANGE_INVALID
        - PASSWORD_POLICY_VIOLATION
        - RESOURCE_NOT_FOUND
        - IMPORTANT_DATE_CONFLICT
        - WATCHLIST_NAME_CONFLICT
        - WATCHLIST_MEMBER_CONFLICT
        - WATCHLIST_CAPACITY_EXCEEDED
        - RATE_LIMITED
        - EXTERNAL_DEPENDENCY_UNAVAILABLE
        - INTERNAL_SERVER_ERROR
    ErrorDetail:
      type: object
      required: [field, code, message]
      properties:
        field: { type: [string, 'null'] }
        code: { type: string }
        message: { type: string }
    Pagination:
      type: object
      required: [limit, offset, total, has_more]
      properties:
        limit: { type: integer, minimum: 1 }
        offset: { type: integer, minimum: 0 }
        total: { type: integer, minimum: 0 }
        has_more: { type: boolean }
    SuccessResponseBase:
      type: object
      required: [code, message, errors, request_id, timestamp]
      properties:
        code: { type: integer, const: 0 }
        message: { type: string, const: '' }
        errors:
          type: array
          maxItems: 0
          items: { $ref: '#/components/schemas/ErrorDetail' }
        request_id: { type: string, minLength: 1 }
        timestamp: { type: string, format: date-time }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, data, errors, request_id, timestamp]
      properties:
        code: { $ref: '#/components/schemas/BusinessErrorCode' }
        message: { type: string, minLength: 1 }
        data: { type: 'null' }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
        request_id: { type: string, minLength: 1 }
        timestamp: { type: string, format: date-time }
    BadRequestErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, enum: [200001, 200002, 200003] }
    UnauthorizedErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, enum: [100001, 100002] }
    ForbiddenErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, const: 100003 }
    NotFoundErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, const: 300001 }
    ConflictErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, enum: [400001, 400002, 400003, 400004] }
    RateLimitedErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, const: 500001 }
...
    InternalServerErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            code: { type: integer, const: 900001 }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 64 }
        password: { type: string, minLength: 1, maxLength: 128 }
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string, minLength: 1, maxLength: 128 }
        new_password: { type: string, minLength: 12, maxLength: 128 }
    User:
      type: object
      required: [user_id, username, display_name]
      properties:
        user_id: { type: string, format: uuid }
        username: { type: string }
        display_name: { type: string }
    AuthSession:
      type: object
      required: [user, csrf_token, expires_at]
      properties:
        user: { $ref: '#/components/schemas/User' }
        csrf_token: { type: string }
        expires_at: { type: string, format: date-time }
    ImportantDateInput:
      type: object
      required: [event_date, title]
      properties:
        event_date: { type: string, format: date }
        title: { type: string, minLength: 1, maxLength: 120 }
        notes: { type: [string, 'null'], maxLength: 1000 }
    ImportantDate:
      allOf:
        - $ref: '#/components/schemas/ImportantDateInput'
        - type: object
          required: [important_date_id, created_at, updated_at]
          properties:
            important_date_id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    CalendarDay:
      type: object
      required: [date, market_code, market_status, important_dates]
      properties:
        date: { type: string, format: date }
        market_code: { type: string, enum: [CN-S] }
        market_status: { type: string, enum: [OPEN, CLOSED, UNKNOWN] }
        previous_open_date: { type: [string, 'null'], format: date }
        calendar_updated_at: { type: [string, 'null'], format: date-time }
        important_dates:
          type: array
          items: { $ref: '#/components/schemas/ImportantDate' }
    StockSummary:
      type: object
      required: [stock_id, market_code, venue_code, security_code, name, listing_status]
      properties:
        stock_id: { type: string, format: uuid }
        market_code: { type: string }
        venue_code: { type: string }
        security_code: { type: string }
        name: { type: string }
        listing_status: { type: string }
...
    BrokerRecommendation:
      type: object
      required: [recommendation_id, recommendation_month, broker_name, stock, updated_at]
      properties:
        recommendation_id: { type: string, format: uuid }
        recommendation_month:
          type: string
          format: date
          pattern: '^\d{4}-\d{2}-01$'
          description: 推荐月份，以当月首日表示
        broker_name: { type: string }
        stock: { $ref: '#/components/schemas/StockSummary' }
        updated_at: { type: string, format: date-time }
...
    BrokerRecommendationPageData:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/BrokerRecommendation' }
        pagination: { $ref: '#/components/schemas/Pagination' }
    AuthSessionResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/AuthSession' }
    CalendarResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/CalendarDay' }
    ImportantDateResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/ImportantDate' }
...
    BrokerRecommendationListResponse:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SuccessResponseBase'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/BrokerRecommendationPageData' }
...